DzRuntimePluginAction renamed to DzBridgeAction
+ also: BridgeDialog moved to DzBridgeAction->m_bridgeDialog
